9C-431 Fix Randomly failing test.
There is one test in CollectionProcessImplSpec that has been failing in
some of the automatic builds in Travis. I think that the error is caused
due to the line that defines the variable newPosition, because it
expects that the new position is different from the previous one.

This is a quick fix to stop the random error happening again.