9C-431 Repair case in
Apart from the error with the random generator, there was also a
failure with the implementation of the method. The case in which
newposition was equal to the old one should not have given errors.

The mistake was in the use of the `if` construct within a For loop,
which defaulted to a failure. We have repaired that as well.